Changes for a 64bit build.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@1680 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ ecc/Elkhound/grampar.cc
@@ -1210,7 +1210,7 @@ GrammarAST *parseGrammarFile(rostring origFname, bool useML)
   // open input file
   SourceManager SM;
   FileManager FM;
-  if (fname == NULL) {
+  if (fname == (int)NULL) {
     // Read from stdin.
     llvm::MemoryBuffer *SB = llvm::MemoryBuffer::getSTDIN();
